+++ src/main/java/org/dainn/dainninventory/entity/OrderDetailEntity.java
@@ -7,6 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -16,13 +18,13 @@ import lombok.Setter;
 public class OrderDetailEntity extends BaseEntity{
     @Column(name = "quantity", nullable = false)
     @Min(1)
-    private int quantity;
+    private Integer quantity;
 
-    @Column(name = "price", nullable = false)
-    private double price;
+    @Column(name = "price", nullable = false, precision = 10)
+    private BigDecimal price;
 
-    @Column(name = "total", nullable = false)
-    private double total;
+    @Column(name = "total", nullable = false, precision = 10)
+    private BigDecimal total;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id")
